==> code/notebooks/04.Movielens_pipeline.ipynb <==
import os
import json
!pip install kaggle

# ========= CONFIG =========
target_directory = os.path.join('..', '..', 'data', 'raw', 'movielens')
os.makedirs(target_directory, exist_ok=True)

# ========= SAVE KAGGLE TOKEN =========
kaggle_token = {
    "username": "nicovansteen",
    "key": "*****"
}

os.makedirs(os.path.expanduser("~/.kaggle"), exist_ok=True)
with open(os.path.expanduser("~/.kaggle/kaggle.json"), "w") as f:
    json.dump(kaggle_token, f)

# ========= DOWNLOAD & EXTRACT =========
!kaggle datasets download -d grouplens/movielens-20m-dataset -p "$target_directory"
!unzip -q "$target_directory/movielens-20m-dataset.zip" -d "$target_directory"
os.remove(f"{target_directory}/movielens-20m-dataset.zip")
